+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Made non-default options lowercase when prompting for transaction confirmation (#843)
 * [BREAKING] Updated keystore to accept arbitrarily large public keys (#833).
 * Added Examples to Mdbook for Web Client (#850).
+* Added account code to `miden account --show` command (#835).
 
 ## 0.8.2 (TBD)